requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Machine Learning
scikit-learn==1.4.0
pandas==2.2.0
joblib==1.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Streamlit
streamlit==1.31.0

# HTTP Client
requests>=2.31.0
